Add types for Damas-Milner typing.
This commit reorganizes the type system to be compatible with
Damas-Milner typing. In other words, the type system is now a formal
type language that includes type variables and type literals.

This commit also includes some tools for working with the new type
system, such as a unification algorithm. The current implementation of
unification is a slow (exponential) recursive algorithm, but this can be
improved later while preserving the same user interface.