Move occurs check from %in% to %occurs_in%.
This commit makes the %in% generic behave more like the base R %in%
function, meaning it checks for a Term in a vector-like object.

Previously, the %in% generic checked for a Variable anywhere in a Term,
which is useful but different from how base R %in% behaves. Now this
check is available via %occurs_in% instead.